tubeup.TubeUp: Fix youtube-dl downloading progress bar
Before this, youtube-dl downloading progress bar was not flush the line
properly.

This commit fix that problem by adding "erase to end of line" escape
sequence to the progress bar string.

And also add a newline character to a string, so that string not
'appended' to the downloading progressbar string.