Add --use-download-archive flag
Add --use-download-archive flag to tubeup.

`--use-download-archive` flag:
Record the video url to the download archive.
This will download only videos not listed in
the archive file. Record the IDs of all
downloaded videos in it.

By using this flag tubeup will use `~/.tubeup/.ytdlarchive`
as `download-archive` argument to `youtube_dl`