Ability, from an existing info-file, to upload existing associated media files.
If `tubeup` is passed the path of info.json files instead of URL,
then it will parse them and, if `--output` was set correctly (same value as previous `yt-dlp`)
then recover the basename of the local files associated with this video.

If used intelligently, especially in conjunction with:
--ignore-existing-item # needed, for now, to avoid the download-archive codepath
--yt=--format=...      # use the same download format
--yt=--no-overwrites
... then existing files may be uploaded without having been re-downloaded/overwritten.

Usage isn't foolproof. Use at your own risks.